fix spacing

+++ docx-core/src/reader/run_property.rs
@@ -31,7 +31,7 @@ impl ElementReader for RunProperty {
                         XMLElement::Size => rp = rp.size(usize::from_str(&attributes[0].value)?),
                         XMLElement::Spacing => {
                             if let Some(v) = read_val(&attributes) {
-                                rp = rp.spacing(u32::from_str(&v)?)
+                                rp = rp.spacing(f32::from_str(&v)? as i32)
                             }
                         }
                         XMLElement::Underline => rp = rp.underline(&attributes[0].value.clone()),
